+++ packages/native_pdf_renderer/CHANGELOG.md
@@ -1,3 +1,19 @@
+## 1.4.2
+
+* Fixed not correctly filling background Color on IOS
+
+## 1.4.1
+
+* Resolve supports Flutter v1.7
+
+## 1.4.0+1
+
+* Hotfix
+
+## 1.4.0
+
+* Now `format` and `backgroundColor` options for image rendering works on IOS.
+
 ## 1.3.0+1
 
 * Hotfix